+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pre-narrowing is now always enabled for `consult`.
 - `wal-partial-recall-reclaim` can now be forced by calling it with
   `C-u`.
+- `wdb-{faraway,nearby}` now default to
+  `display-buffer-use-some-window`.
 
 ## [v2.1.0]